Update main.py
Added try/except. Checks if an entry exists for distances[end][begin] i.e. (checks if we have computed the reverse of this journey). If there is, takes its value and applies it for the distances[begin][end], and takes the existing paths[end][begin] entry and sets it as the paths[begin][end] entry